+++ NEWS.md
@@ -1,4 +1,12 @@
 
+# pkgcache 1.0.4
+
+* Fix handling of Bioconducor versions and repositories, see
+  README for the details.
+
+* Now different versions of pkgcache, that potentially have different
+  metadata format, can share the same metadata cache directory.
+
 # pkgcache 1.0.3
 
 * Fix concurrency issues when the async API is used multiple times in the